Step 2: Categories Completed with Middleware Security 05JUNE2018 3:54 PM

+++ resources/views/admin/posts/index.blade.php
@@ -10,15 +10,15 @@
                         <div class="card">
                             <div class="header">
 @if(session('update'))
-    <div class="p-3 mb-2 bg-secondary text-white text-center">{{session('update')}}</div>
+    <div class="alert alert-info  text-white text-center" role="alert">{{session('update')}}</div>
 @endif
 @if(session('success'))
-    <div class="p-3 mb-2 bg-secondary text-white text-center">{{session('success')}}</div>
+    <div class="alert alert-success  text-white text-center" role="alert">{{session('success')}}</div>
 @endif
 @if(session('delete'))
-    <div class="p-3 mb-2 bg-secondary text-white text-center">{{session('delete')}}</div>
+    <div class="alert alert-danger  text-white text-center" role="alert">{{session('delete')}}</div>
 @endif
-                            <a class="btn pull-right" href="{{route('posts.create')}}">Add New Post</a>
+                            <a class="btn btn-success pull-right" href="{{route('posts.create')}}">Add New Post</a>
 
                                 <h4 class="title">Posts</h4>
                                 
@@ -41,19 +41,16 @@
                                         <tr>
                                         <td>{{$post->id ? $post->id : 'Not Available'}}</td>
                                         <td><img width=50 height=50 class=" img-thumbnail" src="{{$post->photo ? $post->photo->path : '/images/not_available.jpg'}}" alt=""></td>
-                                        <td>{{$post->title ? str_limit($post->title, 50) : 'Not Available'}}</td>
+                                        <td><a href="{{route('posts.edit',$post->id)}}">{{$post->title ? str_limit($post->title, 50) : 'Not Available'}}</a></td>
                                         <td>{{$post->created_at ? $post->created_at->format('d M Y') : 'Not Available'}}</td>
                                         <td><a class="btn btn-info" href="{{route('posts.edit',$post->id)}}">Edit</a></td>                                     
-                                        <td>
+                                        <td onclick="return confirm('Are you sure you want to delete this item?');">
                                             {!!Form::open(['method'=>'DELETE','action'=>['PostController@destroy',$post->id],'files'=>true])!!}
-                                         
                                             <div class='form-group'>
                                             {!! Form::submit('Delete',['class'=>'btn btn-danger']) !!}
                                             </div>
                                             {!! Form::close() !!}
-                                        
-                                            
-                                        </td>                                     
+                                           </td>                                     
                                         
                                     </tr>
 @endforeach
